pr feedback

+++ packages/world/src/utils/InventoryUtils.sol
@@ -86,12 +86,12 @@ library InventoryUtils {
     return slot;
   }
 
-  function addObject(EntityId owner, ObjectTypeId objectType, uint16 amount) public {
+  function addObject(EntityId owner, ObjectTypeId objectType, uint128 amount) public {
     require(amount > 0, "Amount must be greater than 0");
     uint16 stackable = ObjectTypeMetadata._getStackable(objectType);
     require(stackable > 0, "Object type cannot be added to inventory");
 
-    uint16 remaining = amount;
+    uint128 remaining = amount;
 
     // First, find and fill existing slots for this object type
     uint256 numTypeSlots = InventoryTypeSlots._length(owner, objectType);
@@ -105,7 +105,7 @@ library InventoryUtils {
       }
 
       uint16 canAdd = stackable - currentAmount;
-      uint16 toAdd = remaining < canAdd ? remaining : canAdd;
+      uint16 toAdd = remaining < canAdd ? uint16(remaining) : canAdd;
       InventorySlot._setAmount(owner, slot, currentAmount + toAdd);
       remaining -= toAdd;
     }
@@ -113,7 +113,7 @@ library InventoryUtils {
     // If we still have objects to add, use empty slots
     while (remaining > 0) {
       uint16 slot = _useEmptySlot(owner);
-      uint16 toAdd = remaining < stackable ? remaining : stackable;
+      uint16 toAdd = remaining < stackable ? uint16(remaining) : stackable;
 
       Inventory._push(owner, slot);
       InventorySlot._setObjectType(owner, slot, objectType);
